filenames: rename files to use camelCase convention
