+++ package.json
@@ -29,7 +29,7 @@
   },
   "scripts": {
     "dev": "node src/index.js",
-    "build": "webpack --config webpack/config.dev.babel.js --hide-modules",
+    "build": "webpack --config webpack/config-dev.babel.js --hide-modules",
     "precommit": "npm run lint",
     "lint": "standard --verbose | snazzy"
   }
